require one day before reveal
